Move normandy client to normandy app fixes #1597
