+++ app/experimenter/normandy/__init__.py
@@ -1,0 +1,1 @@
+from experimenter.normandy.client import *  # noqa
