+++ app/pytest.ini
@@ -4,8 +4,3 @@ DJANGO_SETTINGS_MODULE = experimenter.settings
 # -- recommended but optional:
 addopts = -p no:warnings
 python_files = tests.py test_*.py *_tests.py
-testpaths =
-    experimenter/base/tests/
-    experimenter/openidc/tests/
-    experimenter/experiments/tests/
-    experimenter/notifications/tests/
